requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
networkx>=3.1
torch-geometric>=2.3.0
yfinance>=0.2.28
ta>=0.10.2
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.42.0
mlflow>=2.7.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
jupyter>=1.0.0
ipywidgets>=8.0.0
rich>=13.0.0
typer>=0.9.0
